Adds wait time for 'yank line' when running remote shell
If 'yank line' is tried when running in a remote shell (ssh) with
latency, the result will not be good.
I consistently yanked only the *last character* in the command.
This is due to latency: yank line feature works faster than the cursor
is moved in the remote shell.

To fix this issue, if remote shell is detected (ssh or mosh) wait time
is added to compensate for network latency.
Wait time of '0.2s' seemed to give good results, but at the end, wait of
'0.4s' is set. This should be more robust and allow for even greater
latency (I only wonder how could anyone work with that).