Bugfix: line yank doesn't work when in last pane line
When the current command line is the last/bottom line in the scrollback
buffer, then jumping to the bottom of the buffer:
- stays in the current line
- goes to the start of current line

This is, of course, not desired bc we get the wrong result.

The fix is to just 'down' movement in copy mode. The down movement
preserves cursor column.

Fixes #11.