Bugfix: copy line binding never yanks newlines
This addresses a tmux/bash issue that occurs for commands with multiple
lines. When 'C-e' is pressed in a command, this confuses tmux so
newlines are yanked in copy-mode.

When a user wants to line-yank multiple-line command he is better off
when newlines aren't yanked.
This commit reliably enables this.

Fixes #13