scripts: helpers.sh: use primary clipboard on wayland with mouse
Follow the same logic as in X and use the primary clipboard for Wayland
sessions when mouse support is on.

Note a new parameter (@yank_wayland_mouse), which is off by default, was
introduced so that the old behavior is maintained. That means that users
need to set the following in their tmux.conf:

`set -g @yank_wayland_mouse 'true'`

to enable the above behavior.

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>